Добавляются ответственные в combobox на странице AddUser. Много багов, фиксим)

+++ Pages/AddUser.xaml.cs
@@ -27,10 +27,20 @@ namespace Documents_Тепляков.Pages
         public AddUser(User User = null)
         {
             InitializeComponent();
-            if(User != null)
+            AddUser_cB();
+            if (User != null)
             {
                 this.User = User;
-                tb_new_user.Text = this.User.user;
+                cb_user.Text = this.User.user;
+            }
+        }
+
+        public void AddUser_cB()
+        {
+            cb_user.Items.Clear();
+            foreach (UserContext user in MainWindow.init.AllUsers)
+            {
+                cb_user.Items.Add(user.user);
             }
         }
 
@@ -53,26 +63,38 @@ namespace Documents_Тепляков.Pages
                 newUser.Save();
                 MessageBox.Show("Ответственный добавлен");
             }
-            else
+            MainWindow.init.AllUsers = new UserContext().AllUsers();
+            AddUser_cB();
+        }
+
+        private void DeleteUser(object sender, RoutedEventArgs e)
+        {
+            if(tb_new_user.Text != "")
+            {
+                UserContext.Delete();
+                MainWindow.init.AllUsers = new UserContext().AllUsers();
+                AddUser_cB();
+            }
+            else MessageBox.Show("Выбирите ответственного для удаления");
+        }
+
+        private void UpdateUser(object sender, RoutedEventArgs e)
+        {
+            if (tb_new_user.Text != "")
             {
                 UserContext newUser = new UserContext();
                 newUser.user = tb_new_user.Text;
                 newUser.Save(true);
                 MessageBox.Show("Ответственный изменен");
+                AddUser_cB();
             }
-            MainWindow.init.AllUsers = new UserContext().AllUsers();
-            MainWindow.init.OpenPage(MainWindow.pages.main);
+            else MessageBox.Show("Выбирите ответственного для изменения");
         }
 
-        private void DeleteUser(object sender, RoutedEventArgs e)
+        private void Users(object sender, SelectionChangedEventArgs e)
         {
-            UserContext.Delete();
-            MainWindow.init.AllUsers = new UserContext().AllUsers();
-        }
-
-        private void UpdateUser(object sender, RoutedEventArgs e)
-        {
-
+            tb_new_user.Text = cb_user.SelectedItem.ToString();
+            return;
         }
     }
 }
